refactor(items): View items moved to components
